files/requirements.txt
fdk>=0.1.54
requests
oracle-ads
